[BUGFIX] Hide the preview button for sysfolders in list/page layout module
In the list view of a sysfolder, users could still see the page preview
button. The list view is now using the same logic as the page module
and the EditDocumentController, that allows for excluding some doktypes
from showing the preview button.

By default, the page types Recycler, Sysfolder and Spacer are excluded.
This behaviour can be overridden by setting the already available
pageTSconfig "TCEMAIN.preview.disableButtonForDokType".

Resolves: #55300
Releases: master, 9.5
Change-Id: Iad7c651e33928b3ababe84c86222291e0bb6c3cc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63860
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>