[BUGFIX] Use parent page ID as PID for suggest fields
TCA select fields have the capabilities for replacing ###CURRENT_PID###
with the parent page ID of the current record. This should always be
<table>.pid. In case of pages the wrong ID was used.

This was now changed to always use the parent page ID indepentent of
the current record. In case there is no parent page row, 0 is given.

Resolves: #90645
Releases: master, 9.5
Change-Id: I8bc04891a009011f4c0d7fd9e59ab306d650b4d2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63587
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>