[BUGFIX] Respect editPanel.onlyCurrentPid = 1
When accessing the editPanel, the option "onlyCurrentPid" allows to also
check the permissions of the editable record to verify if the user
is allowed to edit it.

When checking for CONTENT_EDIT permissions, the "onlyCurrentPid" is also
evaluated now, having "mayEdit" set to false by default, before evaluating
the condition further.

Resolves: #43429
Releases: master, 9.5
Change-Id: I47ed7a7b8c0d1c750cd5b13b485ed14b3aced2dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63616
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Siepmann <coding@daniel-siepmann.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>