[BUGFIX] Show correct context menu for file mounts
It must not be possible to perform rename/copy/cut/delete actions on
a file mount and this is already respected in the FilemountsProvider
context menu and used in the filelist svg tree.

However, the docheader section in the top right corner of the filelist
module showed the wrong context menu for file mounts, which made it
possible for editors to perform these actions.

Resolves: #90677
Releases: master, 9.5
Change-Id: Iac10b18a4fa3249de9b9c29a0102b3832c933841
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63790
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Haupt <mail@danielhaupt.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Haupt <mail@danielhaupt.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>