[TASK] Streamline BackendLayout API class
This change moves the ConditionMatching back to the BackendLayoutView.

The main goal is to make BackendLayout more of an entity class again,
whereas BackendLayoutView should be the process to build the BackendLayout
objects, and configuration and return them, and act more as a
BackendLayoutResolver - this could / should be renamed in the future.

This change thus reactivates runtime-caching for fetching
a backend layout configuration again.

In addition, the new method "getBackendLayoutForPage()" now
actually returns a "ready-to-use" Backend Layout instance,
which could also used for the Nested Content Grid in the future.

The previous "configurationArray" is now defined as "stucture".

A short-hand function "getUsedColumns()" is added.

Resolves: #90839
Releases: master
Change-Id: Ic6e65317a5faa6d70a181fc3f24b3716a1137c39
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63918
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>