[BUGFIX] Fix thrown exception in history
When a logged in backend user is entering the history/undo functionality
for a certain page, which contains a deleted one, the backend user
authentication will throw an error.

This is due to the fact the pid field is not taken into account when
fetching the pages to check if they are deleted. This pid field will be
used to calculate the page access, but is now done with the value null.

This patch will add the pid field when fetching the subpages which will
be checked if they are deleted.

Resolves: #90963
Releases: master, 9.5
Change-Id: I1b78d4a71fea35b5d55a1d8989e586f01e7d9280
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63989
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>