[BUGFIX] Properly initialize backend user base state in middlewares
Due to the fact that backend user objects have not been initialized
properly in PSR-15 middlewares defining the current in-memory state
for workspaces falls back to the users' default workspace.

Basically `BackendUserAuthentication::groupData['workspace_perms']`
was not initialized which results in the mentioned behavior for
non-admin users.

`fetchGroupData` was split up into public `initializeGroupData`
to be used internally only and protected `enrichGroupData`.

Resolves: #90954
Releases: master, 9.5
Change-Id: I75ad15ac600ca489b9951199482e97bcb54d1778
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64264
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Benni Mack <benni@typo3.org>