[BUGFIX] Explicitly fetch element by ID instead of invoking querySelector
The InlineControlContainer tried to fetch elements via `querySelector()`
previously. This fails once the selector contains a character that has a
specific meaning for the whole selector, e.g. a dot. Having a selector
`#foo-bar-settings.test` instructs the query selector to fetch an
element matching the ID `foo-bar-settings` and the class `test`.

As we handle IDs only, we can safely use `getElementById()` which
doesn't have the previously mentioned implications.

Resolves: #91183
Releases: master
Change-Id: I63d23428ee548140f42e2c551d155ac829a2b596
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64343
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Raphael Zschorsch <rafu1987@gmail.com>
Tested-by: Armin Vieweg <armin@v.ieweg.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Raphael Zschorsch <rafu1987@gmail.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>