[BUGFIX] Make edit record bookmarks appear again
The patch for issue #91210 introduced a side effect that bookmarks
to controllers which are not directly registered as backend modules
do not show up again because the ModuleLoader->checkMod() call does
not return an array anymore.
Bookmarks to records of FormEngine / EditDocumentController
are affected by this and are simply not rendered as bookmark
toolbar items.
The patch applies a hack to the ShortcutRepository to deal with
this. We need a better module / controller access API substituting
TBE_MODULES to deal with this properly. EditDocumentController
does it's own access checks for records, so the patch shouldn't
add security harm.

Releases: master, 10.4
Resolves: #91368
Related: #91210
Change-Id: I0add57326b12182a1c11ee6bd09e61b0f1c204f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64476
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Helmut Hummel <typo3@helhum.io>