[BUGFIX] Bring back record type icons for select items
With the removal of `selicon_field_path` in #87937 also the
automatic record type icon mapping was removed.

As a result the record icon of a select item based on `foreign_table`
is not resolved anymore. In addition, the `selectIcons` list is therefore
no longer displayed.

The previous functionality is now restored.

Resolves: #91302
Relates: #87937
Releases: master
Change-Id: If62f4ba65ef54ec2345131f6c117ce4336e76c4c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64560
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>