[!!!][TASK] Replace bootstrap-datetimepicker
This patch replaces the bootstrap-datetimepicker with flatpickr which
is pretty lightweight.

Since this datetimepicker also has some sanity validation on board, much
custom datetime parsing in FormEngine can and has been removed,
including magic date manipulation by using addition or subtraction.

The year format for momentjs has been changed from `YYYY` to `Y` as well
to support years < 1000.

Currently the datepicker is not able to select a date < 100.
By manual selection, the picker jumps back to 1999 and not to 99 as
expected.

Resolves: #91606
Releases: master
Change-Id: Id723c5200c797019f6de0ed5a4c15b2137a32995
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64833
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>