[TASK] Move EXT:about into EXT:backend
The about system extension is responsible for the log-in screen (if EXT:dashboard
is not installed), and despite having just one single controller, EXT:backend
is depending on the extension / module when EXT:dashboard is not installed.

To help newcomers reduce the number of (from their PoV) extensions
that need to be installed via composer, or need to be upgraded, this package
is moved into EXT:backend.

Resolves: #91888
Releases: master
Change-Id: I3e37ba9d9dbc850d1ae0f2c7b345e3526d35029d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65036
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>