[BUGFIX] Drop obsolete CSS code
This patch removes a lot of either unused or superfluous CSS code.

Some stuff seems to have travelled in time to stay present, e.g. a star
hack for IE or code affecting ExtJS which was removed in v9.

The class `form-field-inputlink-explanation` was in use but has been
removed nevertheless as it had no effect since we have Bootstrap in
place.

Resolves: #91867
Releases: master, 10.4
Change-Id: Idc2dd3fa46d393024abce80b46f796f3de867ee2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65112
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>