[BUGFIX] IRRE: Handle localizable records correctly
Each node in an IRRE container now has a new `data-placeholder-record`
attribute to be able to identify such node w/out accessing CSS classes
used for styling purpose.

Nodes where `data-placeholder-record` has a truthy value are localizable
and are ignored now when the sorting is changed which solves an issue
where such records would have been moved into a different language.

In the same run the localization is fixed as it was tried to iterate an
object, which doesn't work. Alternatively, the object values are fetched
and iterated now.

Resolves: #91942
Resolves: #91938
Releases: master, 10.4
Change-Id: Ic6622eb5d6950893f79d4ae9697d7eaa096d2ef9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65214
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65455