[TASK] Broadcast AjaxDataHandler events
Events sent via AjaxDataHandler are now broadcasted by using the
Broadcast Service. This change allows e.g. the topbar to listen to these
events and trigger a reload.

Broadcasting has the nice side effect that the event is sent to all
browser windows which allows multiple simultaneously opened backends
(that share the same domain) to react on that event.

To ease the integration of event listeners, the event payload sent via
classic `dispatchEvent()` is adjusted to share the same structure with
broadcasted messages.

Resolves: #91947
Releases: master, 10.4
Change-Id: I04321f9ab3db309972890e037a1d3790d9b06f79
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65206
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>