[BUGFIX] Drop "All workspaces" tab from workspace module
When "All workspaces" has been integrated somewhere before v6,
it was advertised as a view to see changes from all workspaces.
This however never worked: When calling the workspace module,
"All workspaces" is empty. If switching to a workspace and then
selecting "All workspaces", the tab shows only the changes of
the selected workspace - identical to the normal workspace tab.
Tracing this back in time, this behavior exists at least since v7.
There is not a single bug report in forge about this.

Since this tab is broken for such a long time without any
report, we assume the functionality is barely needed.

A better solution would be to show the number of existing
changes next to the workspace name in the tab list, so editors
can quickly see which specific workspace needs attention, and
then switch to the workspace in question. This should be done with
another patch since non-admins currently see only one workspace tab
at a time, which needs another preparation patch to fix.

This patch drops the broken "All workspaces" tab for now, preparing
further bugfixes to finally end up with a working solution.

Resolves: #91999
Releases: master, 10.4
Change-Id: I33f3cdba117b68927eb10447b0541d8975ed0a63
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65322
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>