[BUGFIX] Workspace relations in ElementInformationController
When relations are added in a workspace, the 'info' modal
of the parent record does not reflect those. The patch changes
the reference resolving to work with the overlay record uid if
given. This fixes reference record display for inline records
including file references.
There are further issues with the rendered record, those
however are resolved by other core methods and can be
fixed with other patches.

Resolves: #45148
Releases: master, 10.4
Change-Id: I7d32a1218eaa2ed47c6cba27e0e52b54e70ae68f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65453
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Achim Fritz <af@achimfritz.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Achim Fritz <af@achimfritz.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>