[!!!][TASK] Drop TCA [ctrl][thumbnail] and user uc[thumbnailsByDefault]
Setting TCA[$someTable]['ctrl']['thumbnail'] to some image related
column made the list module show attached images as preview.
Until core v8, this has been used for tt_content and has been
dropped for this table because two different fields (images, media)
are used and the setting could not cope with that.
For extensions with own tables, this setting has been used
very seldom. It also partially destroys the list module view.

The patch drops evaluation of this ctrl setting in the list module.
With this gone, the 'thumbnailsByDefault' setting of the user settings
module only affects the file list module. The file list module has
it's own checkbox to toggle image preview rendering, so the setup module
checkbox has little benefit and is removed as well. This additionally
fixes a bug that thumbnail preview rendering in file list module can't
be turned off if the setup module checkbox is set.

Additionally, the patch drops some unused css for the list module,
'typo3-dblist' simply does not exist as class.

Change-Id: If9365b5a26e708cc4d4d57cfcddd728cf97d7811
Resolves: #92118
Related: #79622
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65493
Tested-by: Achim Fritz <af@achimfritz.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Windloff
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Achim Fritz <af@achimfritz.de>
Reviewed-by: Daniel Windloff
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>