[BUGFIX] Apply default values before processing datamap
By creating new records the EditDocumentController did not take the
default values fetched via GET / Post variables into account. This led
to multiple issues when defining columns to be readOnly, disabled or
hideForNonAdmins as the values are then stored in defVals parameter
but never added to the final datamap.

The default values are now being added to the Datahandler prior
to process_datamap call.

Resolves: #92119
Resolves: #88456
Resolves: #75514
Releases: master, 10.4
Change-Id: I0975dfd2417711813e1701247e8b68d8805c890b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65569
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>