[BUGFIX] Evaluate drag elements language in page modules Drag&Drop
The underling JavaScript of the Drag&Drop functionality in the page
module passes the columns language, the element is dropped in, to the
DataHandler. This results in L=-1 elements gets changed to L=0 when
moved around in their column.

This is now fixed by using the language uid of the dragged element
instead of the column if elements language is -1.

Resolves: #87148
Releases: master, 10.4
Change-Id: I6663341daadc228480b6dc8ac935b71f117dbfd9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65608
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>