[BUGFIX] Only fix versioning PID for moved records
Since the removal of "pid=-1" in TYPO3 v10, it is not necessary
to fix the versioning PID for all versioned records anymore.

The checks in BackendUtility::fixVersioningPid() and
PageRepository->fixVersioningPid() are cleaned up, because:

* The only version type that has a DIFFERENT page ID than
its online pendant is the MOVE_POINTER (t3ver_state=4)
* The MOVE_PLACEHOLDER does not contain a t3ver_oid
and is not evaluated ever in these methods
* All other t3ver_state items have the same PID as the live
version, and should not be evaluated.

Resolves: #92314
Releases: master, 10.4
Change-Id: I0551547bd2bb9f469ed3a68396b74c0d4de1a18d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65772
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>