[BUGFIX] Send workspace notifications to selected recipients only
When a users sends a draft version to another stage a modal opens
to let the user decide which users should be notified about this.

This patch ensures that only the selected persons will get an email.
Prior to this only the last user of the list got an e-mail.

Resolves: #92316
Change-Id: Ia4014878f6b30a3773b5d38e01d5379094394c32
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65776
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>