[BUGFIX] Do not read workspace if there is no BE user
Running the PopulatePageSlugs upgrade uses the SlugService
to generate the slugs. SlugService itself uses
BackendUtility::getRecordLocalization() which
does a DB query using a WorkspaceRestriction.
This restriction uses the workspace of the current
BE user. But within the Install Tool upgrade there is no
such user of course.

Add the necessary condition to avoid a PHP type error.

Resolves: #92464
Related: #92209
Releases: master, 10.4
Change-Id: I75b6021b4ef1543740ae83bc951b2c2c5374c8d1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65994
Tested-by: Johannes Kasberger <johannes.kasberger@reelworx.at>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Johannes Kasberger <johannes.kasberger@reelworx.at>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Philipp Parzer <p.parzer@siwa.at>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>