[BUGFIX] Ensure getUniqueId returns unique ids for grid objects
The fluid based page module tried to return a "unique id" for each
"create content element" button using the `getUniqueId` method.
This id however was not unique as it was stored as class property
after the first creation.

This lead to all buttons got the same ID which is according to the spec
https://html.spec.whatwg.org/multipage/dom.html#the-id-attribute not
allowed.

Therefore the `getUniqueId()` method now returns a new "unique id"
on each call.

Resolves: #92629
Releases: master, 10.4
Change-Id: Ie3724423e91eb11a6caf5933c651d34691da477b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66207
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>