[BUGFIX] Fix value of multiple select choices in site configuration
When extending the site configuration with an extension and
use a select box with multiple choices (selectSingleBox, selectCheckBox,
selectMultipleSideBySide), the activated options are stored
as an array in the config.yaml file. But to show the activated
items in the form again correctly they must be stored as in
comma-separated list.

Resolves: #92685
Releases: master, 10.4
Change-Id: I5d0713a65e78a52b786b707ab4019c448e360294
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66256
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>