[BUGFIX] Fix exclusive values in FormEngine's selectMultipleSideBySide
With splitting the FormEngine in smaller parts, a regression has been
introduced in the behavior of selectMultipleSideBySide. This select type
supports "exclusive values" which are mutually exclusively selectable.

However, due to a typo in the fetched (non-existent) data attribute, the
whole behavior was broken. This patch fixes the typo.

Resolves: #92737
Related: #87324
Releases: master, 10.4
Change-Id: Ice5de95c49d81df94ee1a8257ecab3abfe7d21d7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66362
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>