[BUGFIX] Reimplement mod.defLangBinding for fluid-based page module
If the TSconfig setting is turned on, translations of content elements
are bound to the default record in the display. This means that within
each column with content elements any translation found for exactly
the shown default content element will be shown in the language
column next to the translation.

This feature has been forgotten during the rewrite and is now readded.

Resolves: #92482
Releases: master, 10.4
Change-Id: I408343d3b9a33b3239d1f341f3df36b65d2cd9c8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66673
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>