[BUGFIX] Allow "copy" action even if the user does not have "edit" permission
If the user does not have "edit" permission, she will not be able to see
the "Copy" action in the content menu. Since copy operation is read-
only, this action should be allowed. This fix will allow copying
without "edit" permissions.

The fix is contributed by the University of Basel.

Resolves: #92708
Releases: master, 10.4, 9.5
Change-Id: I018442e148ba1b4e9801a778bd5a00f433b9d0e0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68865
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>