[BUGFIX] Reintroduce the check on options.saveDocNew
Due to a refactoring and "unrolling" of buttons in TYPO3 v9, the TSconfig
options.saveDocNew has been forgotten when adding a button to create a
new record right while editing another one.

Even if the new button nowadays is physically not a "Save and create new"
action in the label, the behaviour is strictly the same as any unchanged
edit will trigger a modal asking whether the changes should be persisted.

Releases: master, 10.4, 9.5
Resolves: #87321
Resolves: #92788
Change-Id: Ic79f7ff06afef0cf9423780eef2d5324e5613664
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66578
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>