[TASK] Deprecate calling AbstractTreeView::getIcon() with records uid
Calling `AbstractTreeView::getIcon()` with the records uid as
first argument is deprecated. This helps to increase type safety
through the core and also properly reflects the expected value
for the parameter, as its name is `$row`.

Side note: There was no benefit in providing just the records uid
since the full record was fetched in that case anyways internally,
but without adding any restrictions or respecting any overlays.

Because the method returns a string in any case, the return type
is furthermore added to the method signature.

Resolves: #92922
Releases: master
Change-Id: I60ccaf17d8244a2a86fb3b9bc377ce19a6a58e69
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66892
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Daniel Haupt <mail@danielhaupt.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Daniel Haupt <mail@danielhaupt.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>