[BUGFIX] Fix dispatching of `formengine.dp.change` event for date fields
With the introduction of a new datetime picker, the listener for the
event `formengine.dp.change` was not attached properly which broke the
validation of updated datetime field. The listener is now attached to
the datetimer field properly.

To avoid introducing jQuery again, the method
`FormEngineValidation.markFieldAsChanged` became aware of non-jQuery
objects.

Resolves: #92913
Related: #91606
Releases: master
Change-Id: I3727f362a7d685fdea9bcdf0f1d3b37eca12d424
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66966
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>