[!!!][BUGFIX] Do not cache a BE users' group list
When a user logs in, the "cache" field "usergroup_cached_list"
is updated in the be_users table.

This value is used so the groups do not need to be recalculated
each time for each user (e.g. in Workspaces when sending
notifications and in Log module).

However, this has serious flaws:
1) If an admin removes a previously active user from a BE
   group, the cached list still contains the old group, until
   the user logs in
2) If a user never logged in before, the list is empty

This change removes this conceptual bug, and builds this
into the GroupResolver to use direct (optimized) DB queries
for these use cases, making the DB field obsolete.

Resolves: #79565
Releases: master
Change-Id: I9d8fd04cc466906d7f8da43887788f48de81c642
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67137
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>