[BUGFIX] FormEngine does not edit workspace delete placeholders
Workspace delete placeholders are just a marker that a
live record should be deleted in live when the delete
placeholder is published. Actual field data does not matter.
Those records should thus never be editable via FormEngine.

With patch, FormEngine throws an exception when a delete
placeholder is edited. Additionally, the module menu language
selector drop down when editing a record is adapted to not
link to delete placeholder workspace translations.

Change-Id: Iaf73777ffb1d01e6fde5bef1eaa4e281a1b46fe1
Resolves: #88153
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67341
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>