[BUGFIX] Hide deleted flexform section in DOM
If a Flexform section gets deleted, the container must be hidden in DOM
and not deleted, as the underlying DataHandler code relies on getting
the field submitted.

This patch is less intrusive as possible to allow a backport to stable
releases. Flexforms needs a major refactoring and should act similar to
IRRE, but this can go into the master branch only.

Resolves: #93061
Releases: master, 10.4
Change-Id: I69dfdd6f2029547d600939f23e21b03abd45042d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67369
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Martin Kutschker <mkutschker-typo3@yahoo.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Martin Kutschker <mkutschker-typo3@yahoo.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>