[BUGFIX] Small issues on the backend login screen
The translation key login.error.capslock is duplicated.
The second instance is now named login.error.capslockStatus
as in the referring Fluid template.

When there are no loginProviders to be chosen from then
the nav region would only be comprised of an empty <ul>.
Do not render the tags for this condition.

Resolves: #93255
Resolves: #93256
Releases: master
Change-Id: I4527eb8863a92366787827b50b82b441512df744
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67370
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Benni Mack <benni@typo3.org>